Add settings 'normalBuffersOnly' and 'excludedFileTypes'.
Also changed the way settings are specified - from individual global
variables to a single dictionary variable.

'normalBuffersOnly' keeps 'nofile' buffers out of the session file.

'excludedFileTypes' is a list of filetypes you want to keep out of the
session file. I don't know how practical this is, but there may be
workflows where it makes sense.